Refine the `static.CallGraph` to add anonymous functions to the callgraph as source nodes
Using static call graph is mostly effective and accurate enough for the TF provider analyisi use case, with one unfortunate that the graph only includes:

- all package-level functions
- all methods of package-level non-parameterized non-interface types
- pointer wrappers (*C).F for source-level methods C.F
- and all functions reachable from them following only static calls

The entry functions of the typed resources in the TF provider are defined as below:

```
func (r VirtualMachineRunCommandResource) Create() sdk.ResourceFunc {
	return sdk.ResourceFunc{
		Timeout: 30 * time.Minute,
		Func: func(ctx context.Context, metadata sdk.ResourceMetaData) error {
                    ...
                },
        }
```

In this case, the actual create function is not part of the call graph as it doesn't belong to any of the above.

This commit refines the `static.CallGraph` to add "Anonymous Functions" (*anonymous functions (from FuncLit,RangeStmt) directly beneath this one*) to the resulting graph. They are acting as source nodes only, as we are not interested about where they are called, as they are only used as the entry functions during the analysis.