Avoid LRO+non-LRO SDK functions causing multiple same APIOperation(s) being recorded.
Using a map to unify multiple ssa functions end up to be the same APIOperation.

E.g. A resource function can reach to DeleteThenPoll(), which in turns can reach to Delete(). Both corresponds to the same delete API operation. In this case, only this operation will be recorded as a result.